requirements.txt
fastapi==0.119.1
uvicorn==0.38.0
openai==2.6.0
python-dotenv==1.1.1
